+++ src/screens/Assets/collectibles/CollectiblesTab.js
@@ -36,7 +36,7 @@ import { COLLECTIBLE, SEND_COLLECTIBLE_FROM_ASSET_FLOW } from 'constants/navigat
 
 // Selectors
 import { useRootSelector, activeAccountAddressSelector } from 'selectors';
-import { useSupportedChains } from 'selectors/archanova';
+import { useSupportedChains } from 'selectors/chains';
 
 // Utils
 import { spacing } from 'utils/variables';
